Members App Incorporated
A Member application was added from another project that handles user registration and authentication. This was added to demonstrate JSONWebToken authentication, for API endpoint restrictions/permissions control. Many files were altered for integration.

The code isn't functioning as desired, mainly due to SSL requirements not being met on my local machine, running the dev server. Although not complete, it's being pushed to the production server in order to test functionality with the SSL requirements being fulfilled. A few more pushed may follow for debugging purposes. It's not ideal, but the work has to be done.